// frontend/src/interfaces/school.interface.ts
export interface SchoolInterface {
    id : number
    username : String
    phone_number : String
    email : String
    is_active : Boolean
    created_date : Date
    updated_date : Date
    avatar: string
    is_admin: Boolean
    name : String
    slug: string
    stablished_year: Number
    created: Date
    updated: Date
    city: String
    region: Number
    capacity: Number
    manager: String
    teacher: string
    office_manager: string
    }